+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
-## 0.1.1
+## 0.1.2 (Unreleased)
+
+FEATURES:
+
+- operator/resource: a resource can include now a Init function to pre-validate the input.
+- operator/resource: you can now add a 'required' tag to any field in the Resource to make it mandatory.
+
+## 0.1.1 (January 25, 2021)
+
+FEATURES:
 
 - Use Boltdb to store the state